add ghost box with mario

+++ 05-SceneManager/Map.cpp
@@ -20,8 +20,8 @@ CMap::~CMap()
 
 void CMap::Render()
 {
-	int FirstColumn = floor(CamX / TILE_WIDTH);
-	int LastColumn = ceil((CamX + CGame::GetInstance()->GetScreenWidth()) / TILE_WIDTH);
+	int FirstColumn = int(floor(CamX / TILE_WIDTH));
+	int LastColumn = int(ceil((CamX + CGame::GetInstance()->GetScreenWidth()) / TILE_WIDTH));
 	if (LastColumn >= TotalColumnsOfMap)
 		LastColumn = TotalColumnsOfMap - 1;
 	int d = 0;
@@ -31,7 +31,7 @@ void CMap::Render()
 			int index = TileMap[CurrentRow][CurrentColumn] - 1;
 			if (index < TotalTiles)
 			{
-				Tiles.at(index)->Draw(CurrentColumn * TILE_WIDTH, CurrentRow * TILE_HEIGHT);
+				Tiles.at(index)->Draw(float(CurrentColumn * TILE_WIDTH), float(CurrentRow * TILE_HEIGHT));
 			}
 		}
 }
